+++ src/dtos/productDtos/productTrendDTO.js
